refactor(ajax): Use simple Observable
- Refactors `ajax` to use a simple `Observable` implementation.
- Moves code to a new ajax-specific folder.
- Adds a lot more documentation to classes and support classes.
- Differentiates between the configuration passed to `ajax` (`AjaxConfig`) and the request values used to make the HTTP request (`AjaxRequest`), as the latter has more required values.
- Ensures that no configuration values are mutated while making the request.
- Ensures there is at least one valid test for `ajax.patch`. The old one was sketchy.
- Adds better comments throughout the code.
- Adds better typing to `ajax` functions.

BREAKING CHANGE: For TypeScript users, `AjaxRequest` is no longer the type that should be explicitly used to create an `ajax`. It is now `AjaxConfig`, although the two types are compatible, only `AjaxConfig` has `progressSubscriber` and `createXHR`.
